requirements.txt
discord.py
gspread
oauth2client
dotenv